fix/header bug on search after update plofile
