+++ app/page.tsx
@@ -1,10 +1,5 @@
-import { Suspense } from "react";
 import { Top } from "./features/top/Top";
 
 export default function Home() {
-  return (
-    <Suspense>
-      <Top />
-    </Suspense>
-  );
+  return <Top />;
 }
